refactor: init shadcn and update tutorial content
